Qt: allow Packet Bytes view to be shrinked
Even if a packet has only one or two lines for the byte view, Qt
autoresizing insists to require a minimum of 5.5 lines (or 7.5 lines for
packets without multiple data sources). Remove this artificial
requirement and allow the packet bytes view to be resized to show
nothing (except for possible data sources tabs).

This makes it easier to fit more packets and details for screenshots.

Change-Id: I3ea997b9effa8292b396dc2ceb2ab1c35cead410
Reviewed-on: https://code.wireshark.org/review/20342
Reviewed-by: Gerald Combs <gerald@wireshark.org>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>