Fix the calculation of the number of levels in the radix tree.
The algorithm being used calculated the number of levels in a
1024-leaf-node tree as being 2, but it's 1 - 0 elements means 0 levels,
1 through 1024 elements means 1 level, 1025 through 1024^2 elements
means 2 levels, etc..  With a count of 1024, the loop would bump the
level count from 0 to 1, and divide the element count by 1024, yielding
1, so the loop would not terminate, and the level count would them go
from 1 to 2 and the element count would go to 0.

This could cause problems if exactly 1024 packets were seen.

Just use an if chain, similar to the one used when adding elements to
the tree.

Bug: 13433
Change-Id: I3eaeaf374bb65b37b38a59e95f77cac6690614ed
Reviewed-on: https://code.wireshark.org/review/20379
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit b7e9582fd72d85373d5223394d55f405edd0eca8)
Reviewed-on: https://code.wireshark.org/review/20380