Rename a data structure to reflect what it represents.
The structure called a pcap_options structure doesn't (only) hold
options for a packet capture; it holds all information about a
particular device, pipe, or socket from which we're capturing.  Rename
it capture_src - and rename all the pointer variables that point to
something of that type to pcap_src from pcap_opts.

Change-Id: I6f54b3a6d387551d071f02522219948398e5286b
Reviewed-on: https://code.wireshark.org/review/20418
Reviewed-by: Guy Harris <guy@alum.mit.edu>