MySQL: Bugfix state handling
A "conversation" is created for MySQL to store "conversation data" including
current state.  The problem is that it's updated every time a packet is dissected
so depending on how packets are viewed, dissection may change.
Fix that by only allowing conversation state to be set on first pass and have
the state cached with each individual pdu, keyed by tvb_raw_offset() (the key
when you have multiple pdus in a TCP packet).

Change-Id: I2c7e38be63501360d709210a509aa0dfa482bfd4
Reviewed-on: https://code.wireshark.org/review/20600
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>