Add bounds checks for references to outhdr_values.
Don't run past the end of the outhdr_values array.

While we're at it, don't make it a global array - it doesn't need to be
global, as its values don't persist from invocation to invocation, and
this would allow use of this routine in multiple threads in parallel.
(Yes, we have other global variables that don't maintain state, but we
should get rid of that, too.)

Change-Id: I7ceed3c9ba6356eddb7d52472e9d53dd08f90029
Reviewed-on: https://code.wireshark.org/review/20770
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Guy Harris <guy@alum.mit.edu>