Distinguish between the two flavors of ITU-T AFIs.
As ISO 8348/X.213 Annex A says:

A.5.2.1.2 Format and allocation of the IDI

A specific combination of IDI format and DSP abstract syntax is
associated with each allocated AFI value, as summarized for use in
individual addresses in Table A.4 (the corresponding AFI values for use
in group addresses is found via Table A.2).  Two AFI values are
associated with each combination that involves a variable-length IDI
format.  In each case, both of the AFI values identify the same
combination of IDI format and DSP abstract syntax.  The numerically
lower AFI value is used when the first significant digit in the IDI is
non-zero.  The numerically greater AFI value is used when the first
significant digit in the IDI is zero.

and A.5.3 talks further about this.

Change-Id: I3fdad04064451b3c891731dc8f107bc57586eb84
Reviewed-on: https://code.wireshark.org/review/20802
Reviewed-by: Guy Harris <guy@alum.mit.edu>