Fix IPv6/IPX statistics generation
g70381f7 switched from a sorted GSList to a wmem_tree.
But the strings "ipx" or "ipv6" must be tested before "ip" to select the
right tap so the sorting does matter.
Use a wmem_list and parse it by reverse order like what was done with
the GSList to ensure proper selection of the tap.

Change-Id: Ic8d291c265e9c83c960d9a2f6713c25e3aa0e60d
Reviewed-on: https://code.wireshark.org/review/20813
Reviewed-by: Michael Mann <mmann78@netscape.net>
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>