Fix some dissection errors.
"XXX supported" for a metric type is a Boolean, not an int; add it as
such.

Add the value of the item without the extra bits.

The length of an address prefix is in semi-octets (nibbles/hex digits),
not in octets.

Change-Id: I642f0dab5030f7609e89f45cf2cff15cd74dfbda
Reviewed-on: https://code.wireshark.org/review/20819
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit 02f74e47ed3e542f4aff9fc6a7a9c44e29be5c7b)
Reviewed-on: https://code.wireshark.org/review/20830