Improve the display of metrics.
Give the value and the "supported" bit names that reflect what they are.

Show both the "supported" bit and the value as bitfields, because that's
what they are.

Show the "supported" bit before the value.

Show the metric value as decimal; there's nothing particularly
hexadecimal about it.

The "supported" bit is 1 if it's *not* supported and 0 if it *is*
supported.

Fetch the byte for the metric within dissect_metric(), rather than in
the call to dissect_metric().

Change-Id: Ief3bb74b273df06e07066ccdede38a2eeedc6db2
Reviewed-on: https://code.wireshark.org/review/20836
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit ebcfc1ba8e85bd37b03f8ae5fd2f48ce07239bca)
Reviewed-on: https://code.wireshark.org/review/20837