Check before setting the length of a protocol item.
Don't assume that 8 + {32-bit unsigned integer} won't overflow.  Use
tvb_ensure_bytes_exist() to ensure that the data in question is present;
it also checks for overflows.

Also, set the length after we've succeeded in dissecting the item - if
we throw an exception, it's because we don't have all the data, so the
tvb_ensure_bytes_exist() would have failed, but this way we at least get
to dissect what data we *do* have.

Change-Id: If27a2e3ed7978c2051ccb2ddba0d498255d0e350
Reviewed-on: https://code.wireshark.org/review/20840
Reviewed-by: Guy Harris <guy@alum.mit.edu>
(cherry picked from commit de032c63e067197cbd9155082d12ffea5aed4c76)
Reviewed-on: https://code.wireshark.org/review/20841