Squelch some compiler warnings.
Cast an ssize_t value to unsigned int before subtracting it from an
unsigned int.  The value is already known to be >= 0 and <= the value
from which we're subtracting it, and that latter value is an unsigned
int, so it's guaranteed to fit into an unsigned int.  (ssize_t can be
bigger than unsigned int.)

Change-Id: I5c4d3c7fa4bf241c9ea72661beca89bc30495a3b
Reviewed-on: https://code.wireshark.org/review/21114
Reviewed-by: Guy Harris <guy@alum.mit.edu>