cmake: fix build without libxml2
When LibXml2 is not found, the LIBXML2_LIBRARIES is set to
"LIBXML2_LIBRARIES-NOTFOUND" rather than an empty string (as other
FindXxx modules do). This results in an error because the variable is
used in target_link_libraries (via epan_LIBS).

As workaround, explicitly clear the variable to ignore the cache entry.

Change-Id: I9e164f9c175b23559ac6bc37f9aca5f41df79d1f
Reviewed-on: https://code.wireshark.org/review/21321
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Reviewed-by: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>
Reviewed-by: Peter Wu <peter@lekensteyn.nl>