Fixed the column info when displaying VP8 sub protocol to RTP.
The RTP header information was no longer displayed, and this is very useful
information. It certainly looked like the VP8 information was supposed to
be extra to the existing info (the dash in the text), and should have been
tacked on the end.

Change-Id: I1cbefac109a7de7fbc78ac2c274071e20261dd4f
Reviewed-on: https://code.wireshark.org/review/21496
Petri-Dish: Alexis La Goutte <alexis.lagoutte@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>