Clean up various things.
Sort the hf_ definitions, and the entries for them, as much as possible
by the order in which they're put into the protocol tree.

Indicate which if statement branches are for pre-OCTO hardware and which
are for OCTO hardware.

Pre-HT PHYs don't have an MCS index; the "MCS index" is, rather, an
index into a table of bit rates, so call it a "rate index".

Fill in some more 802.11 radio pseudoheader fields.

Use proto_item_append_text to indicate the direction in the entry for
the L1 information.

Don't show irrelevant fields, such as the preamble length for HT and
VHT.

Change-Id: Ibd2baff09fd4ec6e4a84dc0ae2e4615631a2d15e
Reviewed-on: https://code.wireshark.org/review/21550
Reviewed-by: Guy Harris <guy@alum.mit.edu>