Fix various field names and types.
Make some names used for multiple fields more stylistically consistent.

If a bitfield has more than one bit, it's not Boolean - the extra bits
mean it has more values than "true" or "false".

If a bitfield has only one bit, and indicates that something is
true/false, or enabled/disabled, or present/absent, or on/off, or
something else obviously Boolean, it's Boolean.

Change-Id: I89d5e214eb5c4e7f72eb59a78757f0f49fec3dbf
Reviewed-on: https://code.wireshark.org/review/21579
Reviewed-by: Guy Harris <guy@alum.mit.edu>