Give fields better names than ".bitN".
".bitN" names 1) don't indicate what the bit actually *means* and 2) run
the risk of collisions if the bit in question has a different meaning in
different bitsets.

Combine the "receive case" and "transmit case" bitsets in some cases
where the only difference between the two sets is that some bits are
present only in one set, but if a bit is present in both sets, it has
the same meaning.

Change-Id: Ie1cb9d076e431b30b64ef05acab39efe38193d33
Reviewed-on: https://code.wireshark.org/review/21581
Reviewed-by: Guy Harris <guy@alum.mit.edu>